feat: rename skip to drop
No idea why we named it skip in the first place, xstream called it drop
too.

BREAKING_CHANGES: `skip` is now `drop`, `last` does not need to be
called any more in a `pipe` call
