+++ README.md
@@ -42,8 +42,8 @@ Currently, the following set of operators is implemented, others might follow. N
 - [`first`](./src/take.ts)
 - [`filter`](./src/filter.ts)
 - [`startWith`](./src/startWith.ts)
-- [`skip`](./src/skip.ts)
-- [`last`](./src/skip.ts)
+- [`drop`](./src/drop.ts)
+- [`last`](./src/drop.ts)
 - [`flatten`](./src/flatten.ts)
 - [`sample`](./src/sample.ts)
 - [`sampleWith`](./src/sample.ts)
